+++ assets/javascripts/kitten/components/structure/cards/project-card/stories.js
@@ -171,46 +171,83 @@ WithVideo.args = {
 }
 WithVideo.argTypes = argTypes
 
+export const PermanentProjects = ({ progressProps, ...args }) => (
+  <>
+    <ProjectCard stretch {...args}>
+      <ProjectCard.Sticker className="k-u-transform-capitalize k-u-hidden@xs-down">
+        Prolongation
+      </ProjectCard.Sticker>
+      <ProjectCard.Title>
+        The Office, la série culte décortiquée par S!CK
+      </ProjectCard.Title>
+      <ProjectCard.Line>
+        <Text size="micro">
+          <Truncate lines={2}>
+            Ne laisser personne pour compte : plus qu’une devise. Un objectif et
+            une valeur de Un objectif et une valeur de Un objectif et une valeur
+            de
+          </Truncate>
+        </Text>
+      </ProjectCard.Line>
+      <ProjectCard.Line>
+        <Text size="micro">
+          par <strong className="k-u-weight-bold">Bidules</strong>
+        </Text>
+      </ProjectCard.Line>
+      <ProjectCard.Line className="k-u-flex k-u-flex-justifyContent-sb k-u-flex-alignItems-baseline">
+        <div>
+          <Text size="big" weight="bold">
+            172
+          </Text>
+          <Text size="tiny"> abonnés</Text>
+        </div>
+        <Text
+          size="tiny"
+          weight="regular"
+          className="k-u-link k-u-link-primary1"
+        >
+          S’abonner
+        </Text>
+      </ProjectCard.Line>
+    </ProjectCard>
 
-export const PermanentProject = ({ progressProps, ...args }) => (
-  <ProjectCard {...args}>
-    <ProjectCard.Sticker className="k-u-transform-capitalize k-u-hidden@xs-down">
-      Prolongation
-    </ProjectCard.Sticker>
-    <ProjectCard.Title>
-      The Office, la série culte décortiquée par S!CK
-    </ProjectCard.Title>
-    <ProjectCard.Line>
-      <Text size="micro">
-        <Truncate lines={2}>Ne laisser personne pour compte : plus qu’une devise. Un objectif et une valeur de Un objectif et une valeur de Un objectif et une valeur de</Truncate>
-      </Text>
-    </ProjectCard.Line>
-    <ProjectCard.Line>
-      <Text size="micro">
-        par{' '}
-        <strong className="k-u-weight-bold">
-          Bidules
-        </strong>
-      </Text>
-    </ProjectCard.Line>
-    <ProjectCard.Line className="k-u-flex k-u-flex-justifyContent-sb k-u-flex-alignItems-baseline">
-      <div>
-        <Text size="big" weight="bold">
-          172
+    <ProjectCard stretch {...args}>
+      <ProjectCard.Sticker className="k-u-transform-capitalize k-u-hidden@xs-down">
+        Prolongation
+      </ProjectCard.Sticker>
+      <ProjectCard.Title>Un deuxième</ProjectCard.Title>
+      <ProjectCard.Line>
+        <Text size="micro">Petite description</Text>
+      </ProjectCard.Line>
+      <ProjectCard.Line>
+        <Text size="micro">
+          par <strong className="k-u-weight-bold">Bidules</strong>
         </Text>
-        <Text size="tiny">
-          {' '}abonnés
+      </ProjectCard.Line>
+      <ProjectCard.Line
+        lastLine
+        className="k-u-flex k-u-flex-justifyContent-sb k-u-flex-alignItems-end"
+      >
+        <div>
+          <Text size="big" weight="bold">
+            172
+          </Text>
+          <Text size="tiny"> abonnés</Text>
+        </div>
+        <Text
+          size="tiny"
+          weight="regular"
+          className="k-u-link k-u-link-primary1"
+        >
+          S’abonner
         </Text>
-      </div>
-      <Text size="tiny" weight="regular" className="k-u-link k-u-link-primary1">
-        S’abonner
-      </Text>
-    </ProjectCard.Line>
-  </ProjectCard>
+      </ProjectCard.Line>
+    </ProjectCard>
+  </>
 )
 
-PermanentProject.args = args
-PermanentProject.argTypes = argTypes
+PermanentProjects.args = args
+PermanentProjects.argTypes = argTypes
 
 export const Loading = args => (
   <ProjectCard {...args}>