+++ src/esm/components/structure/cards/project-card/index.js
@@ -3,7 +3,7 @@ import _objectWithoutPropertiesLoose from "@babel/runtime/helpers/objectWithoutP
 var _excluded = ["backgroundColor", "alt", "imageClassName"],
     _excluded2 = ["children", "className", "href", "imageProps", "status", "sticker", "videoSources", "videoProps", "stretch", "loading"],
     _excluded3 = ["children", "className"],
-    _excluded4 = ["className"],
+    _excluded4 = ["className", "lastLine"],
     _excluded5 = ["className"],
     _excluded6 = ["className"],
     _excluded7 = ["className"],
@@ -98,10 +98,13 @@ ProjectCard.Title = function (_ref2) {
 
 ProjectCard.Line = function (_ref3) {
   var className = _ref3.className,
+      lastLine = _ref3.lastLine,
       props = _objectWithoutPropertiesLoose(_ref3, _excluded4);
 
   return /*#__PURE__*/React.createElement("div", _extends({
-    className: classNames('k-ProjectCard__line', className)
+    className: classNames('k-ProjectCard__line', className, {
+      'k-ProjectCard__line--lastLine': lastLine
+    })
   }, props));
 };