+++ CHANGELOG.md
@@ -6,6 +6,20 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: Utilities: Add `k-u-padding*` classes.
 - Feature: Utilities: Add `k-u-margin-negative*` classes.
+- Feature: `IconBadge`: Add `micro` to the size prop.
+- Fix: `AlertBox`: Improved styles.
+
+## [8.5.0] - 2022-02-15
+
+Features:
+
+- `ProjectCard`: Add `lastLine` prop to `Line` subcomponent for custom bottom alignment.
+- `DocumentManager`: Add `displayContent` prop for manager states.
+- `ColorSelect`: Add component.
+
+Fix:
+
+- `TextInputWithButton`: Fix `disabled` prop priority.
 
 ## [8.4.0] - 2022-02-14